// src/app/cancela/models/cancela-last-register.ts
export interface CancelaLastRegister {
  assinatura: string;
  cancelaId: number;
  cancelaLiberada: boolean;
  chassi: string;
  clienteId: number;
  clienteNome: string;
  erros: string[];
  id: number;
  leituraBruta: string;
  placa: string;
  placaUrl: string;
  sentido: number;
  tipo: number;
  usuarioCnh: string;
  veiculoId: number;
  veiculoSituacao: number;
}